Kyle's middle part is now working

+++ Backend Code/entry.cpp
@@ -3,7 +3,7 @@
 
 Entry::Entry(){}
 
-Entry::Entry(QString fac, int h, int s){
+Entry::Entry(QString fac, double h, int s){
     faculty=fac;
     hours=h;
     students=s;
